@@ -7,13 +7,16 @@ import com.example.nathan.loadtracker.core.database.entities.Load
 import com.example.nathan.loadtracker.core.database.entities.MyObjectBox
 import io.objectbox.Box
 import io.objectbox.BoxStore
+import io.objectbox.kotlin.boxFor
+import io.objectbox.query.Query
+import io.objectbox.query.QueryBuilder
 
 object LoadTrackerDatabase {
 
     private lateinit var boxStore: BoxStore
 
-    private val loadBox: Box<Load> by lazy { boxStore.boxFor(Load::class.java) }
-    private val jobSessionBox: Box<JobSession> by lazy { boxStore.boxFor(JobSession::class.java) }
+    private val loadBox: Box<Load> by lazy { boxStore.boxFor() }
+    private val jobSessionBox: Box<JobSession> by lazy { boxStore.boxFor() }
 
     fun init(context: Context) {
         boxStore = MyObjectBox.builder()
@@ -22,8 +25,8 @@ object LoadTrackerDatabase {
     }
 
     fun getJobSessions(): List<JobSession> = jobSessionBox.all
-    fun getJobSession(title: String): JobSession = jobSessionBox.query().equal(JobSession_.jobTitle, title).build().findFirst() ?: JobSession()
-    fun getLoadsForSession(jobTitle: String): List<Load> = jobSessionBox.query().equal(JobSession_.jobTitle, jobTitle).build().findFirst()?.loads?.toList() ?: emptyList()
+    fun getJobSession(title: String): JobSession = jobSessionBox.query(JobSession_.jobTitle.equal(title)).build().findFirst() ?: JobSession()
+    fun getLoadsForSession(jobTitle: String): List<Load> = jobSessionBox.query(JobSession_.jobTitle.equal(jobTitle)).build().findFirst()?.loads?.toList() ?: emptyList()
     fun deleteJobSession() {
 
     }
